+++ src/game/world/area_kpa/kpa_82/npc.c
@@ -500,11 +500,11 @@ EvtScript N(EVS_Withdraw_Bobomb_Door_0) = {
 };
 
 // special 'script' used only to organize the quiz waves
-#define QS_NEXT(npcID, script)  npcID, (s32) script,
+#define QS_NEXT(npcID, script)  npcID, (intptr_t) script,
 #define QS_WAIT(n)  -2, n,
 #define QS_END      -1,
 
-s32 N(QS_Release_Wave_0)[] = {
+intptr_t N(QS_Release_Wave_0)[] = {
     QS_NEXT(NPC_DarkTroopa_01,  N(EVS_Release_DarkTroopa))
     QS_WAIT(5 * DT)
     QS_NEXT(NPC_Goomba_01,      N(EVS_Release_Goomba))
@@ -521,7 +521,7 @@ s32 N(QS_Release_Wave_0)[] = {
     QS_END
 };
 
-s32 N(QS_Release_Wave_1)[] = {
+intptr_t N(QS_Release_Wave_1)[] = {
     QS_NEXT(NPC_RedShyGuy_01,   N(EVS_Release_RedShyGuy))
     QS_WAIT(5 * DT)
     QS_NEXT(NPC_RedShyGuy_02,   N(EVS_Release_RedShyGuy))
@@ -546,7 +546,7 @@ s32 N(QS_Release_Wave_1)[] = {
     QS_END
 };
 
-s32 N(QS_Release_Wave_2)[] = {
+intptr_t N(QS_Release_Wave_2)[] = {
     QS_NEXT(NPC_RedShyGuy_01,   N(EVS_Release_RedShyGuy_Door_1))
     QS_NEXT(NPC_BlueShyGuy_01,  N(EVS_Release_BlueShyGuy_Door_2))
     QS_NEXT(NPC_Bobomb_01,      N(EVS_Release_Bobomb_Door_0))
@@ -561,7 +561,7 @@ s32 N(QS_Release_Wave_2)[] = {
     QS_END
 };
 
-s32 N(QS_Release_Wave_3)[] = {
+intptr_t N(QS_Release_Wave_3)[] = {
     QS_NEXT(NPC_Goomba_01,      N(EVS_Release_Goomba))
     QS_WAIT(8 * DT)
     QS_NEXT(NPC_KoopaTroopa_01, N(EVS_Release_KoopaTroopa))
@@ -580,7 +580,7 @@ s32 N(QS_Release_Wave_3)[] = {
     QS_END
 };
 
-s32 N(QS_Release_Wave_4)[] = {
+intptr_t N(QS_Release_Wave_4)[] = {
     QS_NEXT(NPC_Goomba_01,      N(EVS_Release_Goomba))
     QS_WAIT(5 * DT)
     QS_NEXT(NPC_KoopaTroopa_01, N(EVS_Release_KoopaTroopa))
@@ -589,7 +589,7 @@ s32 N(QS_Release_Wave_4)[] = {
     QS_END
 };
 
-s32 N(QS_Release_Wave_5)[] = {
+intptr_t N(QS_Release_Wave_5)[] = {
     QS_NEXT(NPC_RedShyGuy_01,   N(EVS_Release_RedShyGuy))
     QS_WAIT(5 * DT)
     QS_NEXT(NPC_RedShyGuy_02,   N(EVS_Release_RedShyGuy))
@@ -612,7 +612,7 @@ s32 N(QS_Release_Wave_5)[] = {
     QS_END
 };
 
-s32 N(QS_Release_Wave_6)[] = {
+intptr_t N(QS_Release_Wave_6)[] = {
     QS_NEXT(NPC_DarkTroopa_01,  N(EVS_Release_DarkTroopa))
     QS_WAIT(10 * DT)
     QS_NEXT(NPC_BlueShyGuy_01,  N(EVS_Release_BlueShyGuy))
@@ -632,7 +632,7 @@ s32 N(QS_Release_Wave_6)[] = {
     QS_END
 };
 
-s32 N(QS_Withdraw_Wave_0)[] = {
+intptr_t N(QS_Withdraw_Wave_0)[] = {
     QS_NEXT(NPC_DarkTroopa_01,  N(EVS_Withdraw_DarkTroopa))
     QS_NEXT(NPC_DarkTroopa_02,  N(EVS_Withdraw_DarkTroopa))
     QS_NEXT(NPC_DarkTroopa_03,  N(EVS_Withdraw_DarkTroopa))
@@ -643,7 +643,7 @@ s32 N(QS_Withdraw_Wave_0)[] = {
     QS_END
 };
 
-s32 N(QS_Withdraw_Wave_1)[] = {
+intptr_t N(QS_Withdraw_Wave_1)[] = {
     QS_NEXT(NPC_RedShyGuy_01,   N(EVS_Withdraw_RedShyGuy))
     QS_NEXT(NPC_RedShyGuy_02,   N(EVS_Withdraw_RedShyGuy))
     QS_NEXT(NPC_RedShyGuy_03,   N(EVS_Withdraw_RedShyGuy))
@@ -658,7 +658,7 @@ s32 N(QS_Withdraw_Wave_1)[] = {
     QS_END
 };
 
-s32 N(QS_Withdraw_Wave_2)[] = {
+intptr_t N(QS_Withdraw_Wave_2)[] = {
     QS_NEXT(NPC_RedShyGuy_01,   N(EVS_Withdraw_RedShyGuy_Door_1))
     QS_NEXT(NPC_RedShyGuy_02,   N(EVS_Withdraw_RedShyGuy_Door_1))
     QS_NEXT(NPC_BlueShyGuy_01,  N(EVS_Withdraw_BlueShyGuy_Door_2))
@@ -670,7 +670,7 @@ s32 N(QS_Withdraw_Wave_2)[] = {
     QS_END
 };
 
-s32 N(QS_Withdraw_Wave_3)[] = {
+intptr_t N(QS_Withdraw_Wave_3)[] = {
     QS_NEXT(NPC_Goomba_01,      N(EVS_Withdraw_Goomba))
     QS_NEXT(NPC_Goomba_02,      N(EVS_Withdraw_Goomba))
     QS_NEXT(NPC_KoopaTroopa_01, N(EVS_Withdraw_KoopaTroopa))
@@ -684,14 +684,14 @@ s32 N(QS_Withdraw_Wave_3)[] = {
     QS_END
 };
 
-s32 N(QS_Withdraw_Wave_4)[] = {
+intptr_t N(QS_Withdraw_Wave_4)[] = {
     QS_NEXT(NPC_Goomba_01,      N(EVS_Withdraw_Goomba))
     QS_NEXT(NPC_KoopaTroopa_01, N(EVS_Withdraw_KoopaTroopa))
     QS_NEXT(NPC_DarkTroopa_01,  N(EVS_Withdraw_DarkTroopa))
     QS_END
 };
 
-s32 N(QS_Withdraw_Wave_5)[] = {
+intptr_t N(QS_Withdraw_Wave_5)[] = {
     QS_NEXT(NPC_RedShyGuy_01,   N(EVS_Withdraw_RedShyGuy))
     QS_NEXT(NPC_RedShyGuy_02,   N(EVS_Withdraw_RedShyGuy))
     QS_NEXT(NPC_GreenShyGuy_01, N(EVS_Withdraw_GreenShyGuy))
@@ -705,7 +705,7 @@ s32 N(QS_Withdraw_Wave_5)[] = {
     QS_END
 };
 
-s32 N(QS_Withdraw_Wave_6)[] = {
+intptr_t N(QS_Withdraw_Wave_6)[] = {
     QS_NEXT(NPC_DarkTroopa_01, N(EVS_Withdraw_DarkTroopa))
     QS_NEXT(NPC_DarkTroopa_02, N(EVS_Withdraw_DarkTroopa))
     QS_NEXT(NPC_DarkTroopa_03, N(EVS_Withdraw_DarkTroopa))
